Scott's take on what the architecture might look like
Pulled Classes into different folders for organization
Moved the calling of the API into the OnGet of the page so it gets called and Stored per page request
Removed Static variables
Modified scope of some variables (I like my variables to be as tightly scoped as possible)
