+++ HelloWorld/Models/Game.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HelloWorld.Models
+{
+    public class Game
+    {
+        public string Id { get; set; }
+        public string Name { get; set; }
+        public string Box_art_url { get; set; }
+
+    }
+}
