+++ HelloWorld/Program.cs
@@ -17,11 +17,7 @@ namespace HelloWorld
             CreateHostBuilder(args).Build().Run();
         }
 
-        //create new API using Auth Token and Client ID
-        public static TwitchAPI api = new TwitchAPI("or94frz4dm83ru11ql2x9jlik2x9hd", "bscma5xlv1ylbfxj8wafj1jmkt2ycs");
-        //create array to hold top ten games
-        public static TwitchAPI.Game[] games = api.GetTopTenGames();
-
+  
         public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
